requirements.txt
absl-py==2.1.0
addict==2.4.0
alabaster==1.0.0
annotated-types==0.7.0
applaunchservices==0.3.0
appnope==0.1.4
arrow==1.3.0
astroid==3.2.4
asttokens==2.4.1
astunparse==1.6.3
atomicwrites==1.4.1
attrs==24.2.0
autopep8==2.0.4
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.8.0
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
boto3==1.35.3
botocore==1.35.3
catalogue==2.0.10
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
comm==0.2.2
confection==0.1.5
ConfigArgParse==1.7
contourpy==1.2.1
cookiecutter==2.6.0
cycler==0.12.1
cymem==2.0.8
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
docstring-to-markdown==0.15
docutils==0.21.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
flake8==7.0.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
grpcio==1.66.0
h5py==3.11.0
huggingface-hub==0.24.6
idna==3.7
imagesize==1.4.1
importlib_metadata==8.4.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.26.0
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
jellyfish==1.1.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.5.0
keyring==25.3.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
more-itertools==10.4.0
mpmath==1.3.0
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
numpydoc==1.8.0
open3d @ file:///Users/keithyao/Desktop/VSCODE/BuildABouquet/BuildABouquet/open3d-0.18.0%2B18a47ef-cp310-cp310-macosx_12_0_universal2.whl#sha256=6744a356fdde83051383df4b002654403496dee1391f330bdf3cae6df4a77aa9
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.11.1
pydantic==2.8.2
pydantic_core==2.20.1
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.2.6
pylint-venv==3.0.3
pyls-spyder==0.4.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreServices==10.3.1
pyobjc-framework-FSEvents==10.3.1
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.14
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
QDarkStyle==3.2.3
qstylizer==0.2.3
QtAwesome==1.3.1
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rope==1.13.0
rpds-py==0.20.0
Rtree==1.3.0
s3transfer==0.10.2
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==5.5.5
spyder-kernels==2.5.2
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.2
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
text-unidecode==1.3
textdistance==4.6.3
thinc==8.2.5
threadpoolctl==3.5.0
three-merge==0.1.1
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.2
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typer==0.12.4
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.4
whatthepatch==1.0.6
wrapt==1.16.0
wurlitzer==3.1.1
yapf==0.40.2
zipp==3.20.0
